+++ src/utils/getTableRows.ts
@@ -2,39 +2,45 @@ import { SummaryTableRow } from "@actions/core/lib/summary";
 import { TestCase } from "@playwright/test/reporter";
 import Convert from "ansi-to-html";
 
-export const getTableRows = (tests: TestCase[]): SummaryTableRow[] => {
+export const getTableRows = (
+  tests: TestCase[],
+  showError: boolean
+): SummaryTableRow[] => {
   const convert = new Convert();
 
-  const tableRows = [
-    [
-      {
-        data: "Test",
-        header: true,
-      },
-      {
-        data: "Status",
-        header: true,
-      },
-      {
-        data: "Duration",
-        header: true,
-      },
-      {
-        data: "Retries",
-        header: true,
-      },
-      {
-        data: "Error",
-        header: true,
-      },
-    ],
+  const tableHeaders = [
+    {
+      data: "Test",
+      header: true,
+    },
+    {
+      data: "Status",
+      header: true,
+    },
+    {
+      data: "Duration",
+      header: true,
+    },
+    {
+      data: "Retries",
+      header: true,
+    },
   ];
+
+  if (showError) {
+    tableHeaders.push({
+      data: "Error",
+      header: true,
+    });
+  }
+
+  const tableRows = [tableHeaders];
 
   for (const test of tests) {
     // Get the last result
     const result = test.results[test.results.length - 1];
 
-    tableRows.push([
+    const tableRow = [
       {
         data: test.title,
         header: false,
@@ -51,14 +57,19 @@ export const getTableRows = (tests: TestCase[]): SummaryTableRow[] => {
         data: `${result.retry}`,
         header: false,
       },
-      {
+    ];
+
+    if (showError) {
+      tableRow.push({
         data:
           result.error && result.error.message
             ? convert.toHtml(result.error.message!)
             : "",
         header: false,
-      },
-    ]);
+      });
+    }
+
+    tableRows.push(tableRow);
   }
 
   return tableRows;